refactor(layout): move flat layout/ files into per-component subfolders
Applies the same folder convention already used in home/ and products/.
Each component with co-located CSS now lives in its own subfolder.
Related files are grouped together:
  Header/    — Header, HeaderShell
  HeaderNav/ — HeaderNav, HeaderNavContext
  MobileNav/ — MobileNav, MobileNavContext, MobileNavTrigger
  Search/    — SearchPanel, SearchTriggerButton
  MegaMenu/, Footer/, Breadcrumbs/, LocaleSwitcher/

Logomark and PageShell stay flat (no CSS, no sub-components).
All imports updated; tsc passes clean.
