+++ src/components/accessories/AccessoriesShell.tsx
@@ -2,7 +2,7 @@ import type { ReactNode } from "react";
 import {
   Breadcrumbs,
   type BreadcrumbItem,
-} from "@/components/layout/Breadcrumbs";
+} from "@/components/layout/Breadcrumbs/Breadcrumbs";
 import { LT_ACCESSORIES } from "@/lib/content/accessories";
 import type { Locale } from "@/lib/content/home";
 import { AccessoriesSideNav } from "./AccessoriesSideNav";
