+++ src/components/layout/Header/HeaderShell.tsx
@@ -9,10 +9,10 @@ import {
   useRef,
   useState,
 } from "react";
-import { SearchPanel } from "./SearchPanel";
-import { HeaderNavProvider, type HeaderNavCtx } from "./HeaderNavContext";
-import { MobileNavProvider, type MobileNavCtx } from "./MobileNavContext";
-import { MobileNav } from "./MobileNav";
+import { SearchPanel } from "../Search/SearchPanel";
+import { HeaderNavProvider, type HeaderNavCtx } from "../HeaderNav/HeaderNavContext";
+import { MobileNavProvider, type MobileNavCtx } from "../MobileNav/MobileNavContext";
+import { MobileNav } from "../MobileNav/MobileNav";
 import type {
   ShellSearch,
   ShellMobileNav,
